+++ src/pages/LandingPage.jsx
@@ -1,5 +1,5 @@
 import Content from "../components1/Content";
-import EventCalendar from "../components/EventCalendar";
+import EventCalendar from "../components1/EventCalendar";
 import Notice from "../components/Notice";
 import Gallery from "../components/Gallery";
 import styles from "../css/page/LandingPage.module.css";
